atomic: move system containers stuff to a separate file
Split the system containers functionalities from atomic.py to a new
file syscontainers.py.
