+++ atomic
@@ -290,7 +290,7 @@ def create_parser(atomic):
         action="store_true",
         help=_("preview the command that %s would execute") % sys.argv[0])
     installp.add_argument("image", help=_("container image"))
-    if Atomic.atomic.OSTREE_PRESENT:
+    if Atomic.syscontainers.OSTREE_PRESENT:
         installp.add_argument("--system", dest="system",
                               action='store_true', default=False,
                               help=_('install a system container'))
@@ -586,7 +586,7 @@ def create_parser(atomic):
     updatep.add_argument("-f", "--force", default=False, dest="force",
                          action="store_true",
                          help=_("remove all containers based on this image"))
-    if Atomic.atomic.OSTREE_PRESENT:
+    if Atomic.syscontainers.OSTREE_PRESENT:
         updatep.add_argument("--set", dest="setvalues",
                              action='append',
                              help=_("Specify a variable in the VARIABLE=VALUE "
